Move error-tree payload types to their process
Since the typed contract, TreeItem, TreeField, the tree and map
configs and the composite XtfErrorVisualizationConfig are specific to
the XTF error visualization, so they move from the generic
Visualization namespace to Geopilot.Pipeline.Processes
.XtfErrorVisualization, next to the process that produces them. The
generic envelope (Visualization<TData>, IVisualization) stays as the
runtime contract for any visualization type.

VisualizationFactory dissolves into ToVisualization() on the
composite config, which keeps the discriminator bound to its payload
in one place. The wire format is unchanged, pinned by the
serialization test.
