+++ src/Geopilot.Pipeline/Processes/XtfErrorVisualization/LogErrorToTreeItemMapper.cs
@@ -1,5 +1,4 @@
-﻿using Geopilot.Pipeline.Visualization;
-using System.Globalization;
+﻿using System.Globalization;
 using System.Text.RegularExpressions;
 
 namespace Geopilot.Pipeline.Processes.XtfErrorVisualization;
