+++ tests/Geopilot.Pipeline.Test/Processes/XtfErrorVisualizationProcessTest.cs
@@ -1,6 +1,5 @@
 ﻿using Geopilot.Pipeline;
 using Geopilot.Pipeline.Processes.XtfErrorVisualization;
-using Geopilot.Pipeline.Visualization;
 
 namespace Geopilot.Pipeline.Test.Processes;
